refactored again so that review meta data is fetched in RatingsAndReviews component. Makes more sense to have here
